Snip v2.8.0
* Snip will download higher resolution album artwork using a more
efficient method.  This patch was provided by Kailan Blanks.  Thank you.
* Added an option to cache Spotify album artwork locally.  If this
option is enabled then it will download and save the artwork locally to
a separate directory, and then each time that track is played it will
use the saved artwork instead of redownloading it.
* Due to the contribution I added an AUTHORS.txt file.
* Split structs and classes into regions.
* I forced the download timeout to 5 seconds so that there's not too
much of a wait between the time the album artwork starts downloading and
the text file is updated.