Fixed crash related to invalid JSON
It appears that something changed on Spotify's end where it now requires
a user-agent set or it will return an empty string.  I added a fake
user-agent and it appears to work now.

As a safeguard against future crashes it will now just save a blank
image if no JSON is returned.