Fixes #11
This should provide a better fix for not being able to write to the file
because it is in use.  It will just do nothing in the case that it can't
be written to.

In the future I may throw this off into its own thread so that it can
try again without blocking the application until it successfully writes.