Automated resgen process
Resgen is now run via a script to automate the entire process.  The
script will need updated manually each time a new resource file is
added.

I also moved the application signing to an external script that is
manually run afterwards.  This should make it so people can check out
Snip, build it, and run it without issues.