Fix for constant redownloading.
MediaPlayer.Update is called very frequently. While this is good most of the time, before this commit the code didn't check if the album art had actually changed or needed redownloading, and just went ahead and redownloaded it. Now, it checks if either the album ID or size has changed, and only then redownloads it.