Added a newline output option
If you use $n it will add a newline. Note: There's nothing in the UI or
anywhere to tell you this. I'll add something in the future to address
this. For now you can just add $n wherever you want, or replace the
separator with it, and it will add a newline. This addresses #69.