VLC enhancements
Snip will now get the window title of VLC using the process filename of
"vlc.exe" instead of looking for "VLC media player" in the title bar.
This fixes #64.

The problem was that the "VLC media player" text changed based on the
user's locale. This should eliminate that... unless the VLC text comes
before the song information.