Version Check Added
Snip now checks to see if it's the latest version. If a new version is
detected then where the version number shows up in the right-click menu
it will display that there is a new version available instead. If you
click that text it will open your web browser to the download page.