Crash fix for songs with /
This fixes #91. Whenever there was a / in a song title it would get
translated to %2f. When accessing Spotify's web API using a title with
%2f in it it would return 404 Not Found. This is what led to the crash.
I've made it so it replaces the / with a space now. I've tested many
songs with /'s in them and they all work fine now.