Sort Spotify search results
In an attempt to fix #101 Snip now walks through the search results and
selects the result with the highest popularity rating.

I'm assuming a couple of different things here.

First, I'm assuming that this value is based on which track is played
the most frequently by Spotify's users.

Second, I'm assuming that the average Spotify user plays the original
album version of a track the most, and not a remixed version, cover
version, or any other version.

I'm sure the results still won't be 100% but given what I have to work
with I feel it should be a slight improvement.