Fixes song title selection if album has same name
When querying Spotify's servers in the past to get the track information
I would use the first result returned. However, if the song title is
also the title of the album then the song title returned may actually be
another track off of that same album. This is also a fix for #101 as
there were multiple concerns within the same issue.

Thanks to tterrag1098 for the suggestion of comparing the window title
to the track artist and track title. This will ensure that the result is
correct.