Fix invalid JSON string error
Fixes #87.

There are two changes here.

1) Spotify's search does not like the search string as all uppercase.
Some songs would return correctly as-is or forced lowercase, but if they
were forced uppercase the server would return a 404. Because of this
I've forced the strings to all lowercase instead.

2) In the event that the JSON returned is invalid or if the server
returns an error, Snip will now fallback and simply display the window
title instead of crashing.

Thanks to FluxMonkey92 for providing a song that allowed me to duplicate
the issue.