Fixes some variables in song titles
Snip replaces certain variables with other variables. $t becomes the
track name for example. However if the song itself contains one of those
varaibles it would replace it when it shouldn't. The interim solution is
to change the variable names to something less likely to be found in a
song title. I've changed all previous $var to $$var. This fixes #129.