Authorization change
I didn't realize this until I saw people using v6.0.0 but for each
individual user there is a new token generated constantly. From
Spotify's eyes this looks like Snip itself is requesting a _lot_ of
different tokens. To combat this the server side aspect that gets the
token now only obtains one token every 3600 seconds and provides that
one token to every user of Snip. This cuts down on the spam that
Spotify's servers see. I've left the Spotify code in Snip to request
every 3600 seconds still as this will cut down on my server's access as
well. It's a win-win.