Authorization changes
Let's see how this goes.

Initially Snip would grab a new token for every user which would cause a
lot of unnecessary tokens to be pulled from Spotify's server.

I changed the behavior of the remote script to cache a single token,
provide that to every Snip user, and then update every 60 minutes.

The code in Snip would still pull from the server every 60 minutes and
just get the cached token. This seemed like a good idea at first. The
problem is that the server may update every 60 minutes offset from when
Snip would update every 60 minutes. This means it's possible for Snip to
download a token one minute before it expires and not fetch another for
60 minutes, causing Snip not to behave correctly.

The solution is still server side. The difference being that I cache the
token still, but I modify the expires time every time Snip downloads it.
This allows Snip to correctly load the expire time and re-fetch it from
the server when it expires.

There may still be some overlap because of latency in downloading the
file, which may cause Snip to miss details from a song if it happens to
fall right at the same time as the track changes, but I expect this to
be a very odd case.