Moved authorization key
I moved this to a separate file which just contains a global string that
can be called anywhere. I then added this file to .gitignore. This makes
it easier on my part so I don't have to manually edit the key out and
put it back everything I want to commit code or release a new version.