Added Spotify metadata caching
If this option is enabled, which it is by default, Snip will download
and store the JSON from Spotify's servers. The next time the same track
is played it will load the data from the cache instead of downloading
it. This will reduce the amount of calls to Spotify's servers.

Also I changed the way Snip saves and loads its settings in the
registry. Now settings will be remembered across all major versions of
Snip.