Additional Spotify player checks
Cleaned up some of the Spotify code. Removed redundant returns.

Key change is that previously Snip was checking for both valid oauth and
csrf tokens and only following that code path if both were set. The
problem is that only one may be set and the other null. Changed check to
OR instead.

This should help me determine if users that are having issues with
Spotify detection actually can connect to the web service or not.