Better implementation of Spotify detection
This commit slightly changes the behavior of Snip when detecting Spotify. API limitations of Windows prevent Process.MainWindowTitle from working if the process isn't in the taskbar. If the process gets minimized to the system tray Process.MainWindowTItle will return 0, regardless of whether there is a title or not. While it is possible to implement a very ugly solution to enumerate through all window handles and tie them to a process the solution in this commit should suffice. All Snip needs to do is see the Spotify window at least once. If you already have Spotify minimzed to the system tray playing a song and you start Snip, you will need to bring Spotify forward at least once and it will work thereafter. If you start Snip first and then Spotify it will work as intended after minimizing Spotify to the system tray. Fixes #378 for the most part. :)